% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bam_stats}
\alias{bam_stats}
\title{Bam stats}
\usage{
bam_stats(fastq_file)
}
\arguments{
\item{fastq_file}{input sample fastq files}
}
\description{
R function to collect our sample stats
}
\examples{
bam_stats(fastq_file)
}
\keyword{BAM}
\keyword{file}

