% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bwa_align}
\alias{bwa_align}
\title{BWA align}
\usage{
bwa_align(config_df_row)
}
\arguments{
\item{sample}{dataframe}
}
\description{
R function to run the BWA "aln" method on our sample dataframe
}
\examples{
bwa_align(sample_data_frame_row)
}
\keyword{BWA}
\keyword{aln}

