% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bwa_samse}
\alias{bwa_samse}
\title{BWA samse}
\usage{
bwa_samse(config_df_row_row)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to run the BWA "aln" method on our sample dataframe
}
\examples{
bwa_samse(sample_data_frame_row)
}
\keyword{BWA}
\keyword{samse}

