% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{collect_stats}
\alias{collect_stats}
\title{collect stats}
\usage{
collect_stats(config_df)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to collect our sample stats
}
\examples{
collect_stats(sample_data_frame_row)
}
\keyword{BAM}
\keyword{file}

