% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_bam}
\alias{filter_bam}
\title{filter BAM}
\usage{
filter_bam(sort_rmdup_bam, map_qualities = c("15", "25", "30"))
}
\arguments{
\item{sort_rmdup_bam}{Input deduped and sorted bam file}
}
\description{
R function to filter BAM files
}
\examples{
filter_bams("input_sort_rmdup.bam")
}
\keyword{BAM}
\keyword{file}

