% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{filter_bams}
\alias{filter_bams}
\title{filter BAM}
\usage{
filter_bams(config_df_row)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to filter BAM files
}
\examples{
filter_bams(sample_data_frame_row)
}
\keyword{BAM}
\keyword{file}

