% Generated by roxygen2 (4.0.2): do not edit by hand
\name{full_alignment}
\alias{full_alignment}
\title{full_alignment}
\usage{
full_alignment(config_file)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to preform full NGS alignment
}
\examples{
full_alignment("config".csv)
}
\keyword{BWA}
\keyword{SAMtools}

