% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read_samples}
\alias{read_samples}
\title{Read Samples}
\usage{
read_samples(config_file)
}
\arguments{
\item{config}{file in csv format}
}
\description{
This function generates a dataframe to use for the NGS alignment
}
\examples{
read_samples(config_file_csv)
}
\keyword{BWA}
\keyword{NGS}

