% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmdup_bam}
\alias{rmdup_bam}
\title{rmdup BAM}
\usage{
rmdup_bam(config_df_row)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to remove BAM file duplicates
}
\examples{
rmdup_bam(sample_data_frame_row)
}
\keyword{SAMtools}

