\name{rNGS-package}
\alias{rNGS-package}
\alias{rNGS}
\docType{package}
\title{
r packge to drive a NGS alignment using BWA and SAMtools
}
\description{
R code to drive a BWA alignment then filter the results with
    samtools. Both programs need to be installed before the analysis is started
}
\details{
\tabular{ll}{
Package: \tab rNGS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-29\cr
License: \tab GPL3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matthew Teasdale

Maintainer: Matthew Teasdale <matthew@teasdale.xyz>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
