% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{sam_to_bam}
\alias{sam_to_bam}
\title{SAM to BAM}
\usage{
sam_to_bam(config_df_row)
}
\arguments{
\item{sample}{dataframe row}
}
\description{
R function to convert from SAM to BAM files
}
\examples{
sam_to_bam(sample_data_frame_row)
}
\keyword{SAMtools}

