% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sort_bam}
\alias{sort_bam}
\title{Sort BAM}
\usage{
sort_bam(input_bam, sorted_bam)
}
\arguments{
\item{input_bam}{Input BAM file}

\item{sorted_bam}{Output sorted BAM file}
}
\description{
R function to sort BAM files
}
\examples{
sort_bam("input.bam", "output_sorted.bam")
}
\keyword{SAMtools}

