% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bwa_align}
\alias{bwa_align}
\title{BWA align}
\usage{
bwa_align(fastq_file, genome, bwa_arguments, sai_file)
}
\arguments{
\item{fastq_file}{Fastq file for alignment}

\item{genome}{BWA indexed genome to use in alignment}

\item{bwa_arguments}{Arguments for bwa}

\item{sai_file}{Output sai file}
}
\description{
R function to run the BWA "aln" method
}
\examples{
bwa_align("fastq.fa", "genome.fa", "-t 4", "ouput.sai")
}
\keyword{BWA}
\keyword{aln}

