% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bwa_samse}
\alias{bwa_samse}
\title{BWA samse}
\usage{
bwa_samse(fastq_file, genome, read_groups, sai_file, sam_file)
}
\arguments{
\item{fastq_file}{Input fastq file}

\item{genome}{Genome for alignment}

\item{read_groups}{Read group information for alignment}

\item{sai_file}{Input sai file}

\item{sam_file}{Output sam file}
}
\description{
R function to run the BWA "samse"
}
\examples{
bwa_samse("fastq.fa", "genome.fa",
"read group string", "input.sai", "output.sam")
}
\keyword{BWA}
\keyword{samse}

