% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rmdup_bam}
\alias{rmdup_bam}
\title{rmdup BAM}
\usage{
rmdup_bam(input_sorted_bam, output_rmdup_bam)
}
\arguments{
\item{input_sorted_bam}{}

\item{output_rmdup_bam}{}
}
\description{
R function to remove BAM file duplicates
}
\examples{
rmdup_bam("input_sorted.bam", "output_rmdup.bam")
}
\keyword{SAMtools}

