% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{sam_to_bam}
\alias{sam_to_bam}
\title{SAM to BAM}
\usage{
sam_to_bam(sam_file, bam_file)
}
\arguments{
\item{sam_file}{Input SAM file}

\item{bam_file}{Output BAM file}
}
\description{
R function to convert from SAM to BAM files
}
\examples{
sam_to_bam("input.sam", "ouput.bam")
}
\keyword{SAMtools}

